pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "max-agent"
version = "0.1.0"
description = "A Python package for PDF summarization and context embedding for chat agents."
authors = [
    { name = "François REEVES", email = "francois@reevesnco.com" }
]
license = {text = "MIT"}
readme = "readme.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries",
    "Topic :: Text Processing :: Linguistic"
]
dependencies = [
    "fpdf==1.7.2",
    "pdfplumber==0.11.4", 
    "PyYAML==6.0.2",
    "torch",
    "transformers==4.47.0",
    "wordsegment>=1.3.1"
]

[project.urls]
homepage = "https://github.com/Tennisee-data/Max_Agent"

[project.scripts]
max-agent = "max_agent.cli:main"

[project.optional-dependencies]
performance = ["intel_extension_for_pytorch==2.5.0"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"max_agent" = ["*.yaml"]

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311']